El metodo generado por Chatgpt parece funcionar y se agregan las pruebas
